+++ EldritchArcana/Oracle/LifeMystery.cs
@@ -3,46 +3,28 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
 using System.Text;
-using Kingmaker;
-using Kingmaker.AreaLogic;
-using Kingmaker.Assets.UI.LevelUp;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
-using Kingmaker.Blueprints.Classes.Prerequisites;
 using Kingmaker.Blueprints.Classes.Selection;
 using Kingmaker.Blueprints.Classes.Spells;
 using Kingmaker.Blueprints.Facts;
-using Kingmaker.Blueprints.Items;
-using Kingmaker.Blueprints.Items.Equipment;
-using Kingmaker.Blueprints.Root;
-using Kingmaker.Controllers;
-using Kingmaker.Controllers.Combat;
-using Kingmaker.Controllers.Units;
-using Kingmaker.Designers;
 using Kingmaker.Designers.Mechanics.Buffs;
 using Kingmaker.Designers.Mechanics.Facts;
-using Kingmaker.Designers.Mechanics.Recommendations;
 using Kingmaker.ElementsSystem;
 using Kingmaker.EntitySystem.Entities;
 using Kingmaker.EntitySystem.Stats;
 using Kingmaker.Enums;
 using Kingmaker.Enums.Damage;
-using Kingmaker.Localization;
 using Kingmaker.PubSubSystem;
 using Kingmaker.RuleSystem;
 using Kingmaker.RuleSystem.Rules;
 using Kingmaker.RuleSystem.Rules.Abilities;
 using Kingmaker.RuleSystem.Rules.Damage;
 using Kingmaker.UI.Common;
-using Kingmaker.UI.ServiceWindow.CharacterScreen;
 using Kingmaker.UnitLogic;
 using Kingmaker.UnitLogic.Abilities;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
-using Kingmaker.UnitLogic.Abilities.Components;
-using Kingmaker.UnitLogic.Abilities.Components.AreaEffects;
 using Kingmaker.UnitLogic.Abilities.Components.Base;
 using Kingmaker.UnitLogic.Abilities.Components.TargetCheckers;
 using Kingmaker.UnitLogic.ActivatableAbilities;
@@ -50,10 +32,6 @@ using Kingmaker.UnitLogic.Buffs;
 using Kingmaker.UnitLogic.Buffs.Actions;
 using Kingmaker.UnitLogic.Buffs.Blueprints;
 using Kingmaker.UnitLogic.Buffs.Components;
-using Kingmaker.UnitLogic.Class.LevelUp;
-using Kingmaker.UnitLogic.Class.LevelUp.Actions;
-using Kingmaker.UnitLogic.Commands;
-using Kingmaker.UnitLogic.Commands.Base;
 using Kingmaker.UnitLogic.FactLogic;
 using Kingmaker.UnitLogic.Mechanics;
 using Kingmaker.UnitLogic.Mechanics.Actions;
@@ -61,12 +39,6 @@ using Kingmaker.UnitLogic.Mechanics.Components;
 using Kingmaker.UnitLogic.Mechanics.Conditions;
 using Kingmaker.UnitLogic.Parts;
 using Kingmaker.Utility;
-using Kingmaker.View;
-using Kingmaker.Visual.Animation.Kingmaker.Actions;
-using Kingmaker.Visual.Sound;
-using Newtonsoft.Json;
-using UnityEngine;
-using static Kingmaker.RuleSystem.RulebookEvent;
 using static Kingmaker.UnitLogic.ActivatableAbilities.ActivatableAbilityResourceLogic;
 using static Kingmaker.UnitLogic.Commands.Base.UnitCommand;
 
@@ -308,12 +280,14 @@ namespace EldritchArcana
                     deactivated: Helpers.Create<ContextRestoreResource>(c => c.Resource = resource),
                     newRound: Helpers.CreateConditional(
                         Helpers.Create<ContextConditionDistanceToTarget>(c => c.DistanceGreater = 40.Feet()),
-                        new GameAction[] { removeBuff },
-                        new GameAction[] {
-                            Helpers.Create<ContextActionSpawnFx>(c =>
-                                c.PrefabLink = cureLightWounds.GetComponent<AbilitySpawnFx>().PrefabLink),
-                            Helpers.Create<ContextActionTransferDamageToCaster>(c => c.Value = 5)
-                        })));
+                        removeBuff,
+                        Helpers.CreateConditional(
+                            Helpers.Create<ContextConditionHasDamage>(),
+                            new GameAction[] {
+                                Helpers.Create<ContextActionSpawnFx>(c =>
+                                    c.PrefabLink = cureLightWounds.GetComponent<AbilitySpawnFx>().PrefabLink),
+                                Helpers.Create<ContextActionTransferDamageToCaster>(c => c.Value = 5)
+                            }))));
             removeBuff.Buff = buff;
 
             var bondAbility = Helpers.CreateAbility($"{feat.name}Ability", feat.Name, feat.Description,
@@ -579,6 +553,13 @@ namespace EldritchArcana
                 if (Owner.Resources.GetResourceAmount(Resource) == 0) Buff.Remove();
             }
         }
+    }
+
+    public class ContextConditionHasDamage : ContextCondition
+    {
+        protected override bool CheckCondition() => Target.Unit?.Damage > 0;
+
+        protected override string GetConditionCaption() => "Whether the target is damaged";
     }
 
     public class ContextActionTransferDamageToCaster : BuffAction